+++ db_utils/__init__.py
@@ -2,5 +2,5 @@
 
 import os
 
-with open(os.path.join(os.path.dirname(__file__), "VERSION"), 'r') as vf:
+with open(os.path.join(os.path.dirname(__file__), "VERSION"), "r") as vf:
     __version__ = vf.read().strip()