+++ db_utils/command/cleanup.py
@@ -11,13 +11,17 @@ from db_utils.command.base import BaseCommand
 #: query used to get all the data of the already executed filenames
 DELETE_FAILED_MIGRATION = """
 DELETE FROM {table} WHERE filename = %(filename)s
-""".format(table=MIGRATIONS_TABLENAME)
+""".format(
+    table=MIGRATIONS_TABLENAME
+)
 
 #: query used to update the invalid hash of the file when it
 #: was executed manually and the file was edited later on
 UPDATE_HASH = """
 UPDATE {table} SET hashed_content = %(hashed_content)s WHERE filename = %(filename)s
-""".format(table=MIGRATIONS_TABLENAME)
+""".format(
+    table=MIGRATIONS_TABLENAME
+)
 
 
 logger = getLogger(__name__)
@@ -37,26 +41,24 @@ class CleanupCommand(BaseCommand):
         self.validate_migration_table_exists()
         failed_to_run = self.get_executed_migrations(return_success=False)
         if not failed_to_run:
-            logger.warning('There was no failed migration to cleanup')
+            logger.warning("There was no failed migration to cleanup")
             return
 
         for filename in failed_to_run.keys():
-            logger.debug('Found the following filename to cleanup: %s', filename)
+            logger.debug("Found the following filename to cleanup: %s", filename)
 
             if not self.batch_mode and not self.dry_run:
-                should_continue = self.check_input('Should it delete the row for the migration: %s?' % filename)
+                should_continue = self.check_input(
+                    "Should it delete the row for the migration: %s?" % filename
+                )
                 if not should_continue:
-                    print('You chose not to delete the migration')
+                    print("You chose not to delete the migration")
                     return
 
             with self.connection:
                 with self.connection.cursor() as cursor:
-                    self.execute_sql(
-                        cursor,
-                        DELETE_FAILED_MIGRATION,
-                        filename=filename
-                    )
-        logger.info('Finish cleanup of failed migrations')
+                    self.execute_sql(cursor, DELETE_FAILED_MIGRATION, filename=filename)
+        logger.info("Finish cleanup of failed migrations")
 
     def _update_md5(self):
         """
@@ -76,19 +78,21 @@ class CleanupCommand(BaseCommand):
 
             if database_hash is None:
                 # this migration wasn't executed
-                logger.debug('Found the filename %s that was not executed', filename)
+                logger.debug("Found the filename %s that was not executed", filename)
                 continue
 
             if filesystem_hash == database_hash:
                 # the migration was executed and the file wasn't updated
-                logger.debug('Found the filename %s ok', filename)
+                logger.debug("Found the filename %s ok", filename)
                 continue
 
-            logger.warning('Found the filename: %s with different hash', filename)
+            logger.warning("Found the filename: %s with different hash", filename)
             if not (self.batch_mode or self.dry_run):
-                should_cleanup = self.check_input('Found the filename %s with invalid hash. Cleanup?' % filename)
+                should_cleanup = self.check_input(
+                    "Found the filename %s with invalid hash. Cleanup?" % filename
+                )
                 if not should_cleanup:
-                    print('Skip the hash update for file: %s' % filename)
+                    print("Skip the hash update for file: %s" % filename)
                     continue
 
             with self.connection: